Исправление навигации при отсутствии интернета (#82)

+++ frontend/hooks/useHandleLogout.ts
@@ -1,14 +1,20 @@
 import { deleteTokenFromSecureStore } from "@/scripts/jwt";
+import { Router } from "expo-router";
 
-const useHandleLogout = async (router: any) => {
+const useHandleLogout = async (
+  router: Router,
+  deleteTokens: boolean = true
+) => {
   if (router.canDismiss()) router.dismissAll();
   router.replace("/authorize");
 
-  await deleteTokenFromSecureStore("accessToken");
-  await deleteTokenFromSecureStore("refreshToken");
+  if (deleteTokens) {
+    await deleteTokenFromSecureStore("accessToken");
+    await deleteTokenFromSecureStore("refreshToken");
 
-  await deleteTokenFromSecureStore("role");
-  await deleteTokenFromSecureStore("user");
+    await deleteTokenFromSecureStore("role");
+    await deleteTokenFromSecureStore("user");
+  }
 };
 
 export default useHandleLogout;
